# perf(pycodestyle): Initialize Stylist from tokens