--- a/Cargo.toml
+++ b/Cargo.toml
@@ -49,8 +49,6 @@ toml = { version = "0.7.2" }
 
 [profile.release]
 lto = "fat"
-codegen-units = 1
-opt-level = 3
 
 [profile.dev.package.insta]
 opt-level = 3