This PR changes `Stylist` to initialize from tokens instead of re-lexing a string, which, in the worst case, requires re-lexing the whole program. 

This was the reason why `numpy/globals.py` regressed by close to 15% compared to when pydocstyle is disabled. 

## Benchmarks

* `non-logical`: `main` with logical_lines disabled 
* `pr3737`: Base reference with `logical_lines` enabled
* pr3575: This PR with `logical_lines` enabled
* `pr3575-non-logical`: This PR with `logical_lines` disabled

```
group                                      no-logical                             pr3737                                 pr3757                                 pr3757-non-logical
-----                                      ----------                             ------                                 ------                                 -----------
linter/all-rules/large/dataset.py          1.02      8.5±0.01ms     4.8 MB/sec    1.05      8.8±0.09ms     4.6 MB/sec    1.05      8.8±0.06ms     4.6 MB/sec    1.00      8.4±0.16ms     4.8 MB/sec
linter/all-rules/numpy/ctypeslib.py        1.04      2.1±0.01ms     7.8 MB/sec    1.11      2.3±0.06ms     7.3 MB/sec    1.07      2.2±0.01ms     7.5 MB/sec    1.00      2.1±0.01ms     8.1 MB/sec
linter/all-rules/numpy/globals.py          1.00    247.9±3.25µs    11.9 MB/sec    1.10    271.5±1.72µs    10.9 MB/sec    1.00    248.0±2.34µs    11.9 MB/sec    1.00    248.7±5.94µs    11.9 MB/sec
linter/all-rules/pydantic/types.py         1.03      3.7±0.04ms     7.0 MB/sec    1.09      3.9±0.02ms     6.6 MB/sec    1.06      3.8±0.04ms     6.7 MB/sec    1.00      3.6±0.04ms     7.2 MB/sec
linter/default-rules/large/dataset.py      1.00      4.7±0.01ms     8.7 MB/sec    1.07      5.0±0.02ms     8.1 MB/sec    1.08      5.1±0.02ms     8.0 MB/sec    1.00      4.7±0.07ms     8.7 MB/sec
linter/default-rules/numpy/ctypeslib.py    1.00    999.5±5.57µs    16.7 MB/sec    1.09  1088.3±28.69µs    15.3 MB/sec    1.08  1076.3±10.24µs    15.5 MB/sec    1.00  1000.5±12.04µs    16.6 MB/sec
linter/default-rules/numpy/globals.py      1.00    101.0±0.41µs    29.2 MB/sec    1.20    121.7±2.16µs    24.2 MB/sec    1.12    113.5±2.92µs    26.0 MB/sec    1.05    106.4±3.06µs    27.7 MB/sec
linter/default-rules/pydantic/types.py     1.03      2.1±0.01ms    11.9 MB/sec    1.11      2.3±0.02ms    11.0 MB/sec    1.11      2.3±0.04ms    11.0 MB/sec    1.00      2.1±0.04ms    12.2 MB/sec
```

This shows that the performance compared to main is nearly unchanged, but performance improves when enabling `logical_lines` (worst case is now about 10% of regression and no longer 20%)

## CPython Benchmark

This is a end-to-end benchmark comparing the performance between the logical lines implementation of this PR, the logical lines implementation at the start of this stack, and when logical lines is disabled. The results show that

* Reduces the slowdown of enabling logical lines from 1.71 to 1.26
* Caching performance depends on the number of enabled rules... :confused: 
* This stack improves the performance of running all rules by roughly 10%
   
    
### Default Rules - No cache

<pre>
<b>Summary</b>
  &apos;<span style="color:#2AA1B3">./no-logical-main ./crates/ruff/resources/test/cpython/ --no-cache</span>&apos; ran
<span style="color:#26A269"><b>    1.00</b></span> ± <span style="color:#26A269">0.02</span> times faster than &apos;<span style="color:#A347BA">./no-logical-pr ./crates/ruff/resources/test/cpython/ --no-cache</span>&apos;
<span style="color:#26A269"><b>    1.26</b></span> ± <span style="color:#26A269">0.03</span> times faster than &apos;<span style="color:#A347BA">./logical-pr ./crates/ruff/resources/test/cpython/ --no-cache</span>&apos;
<span style="color:#26A269"><b>    1.71</b></span> ± <span style="color:#26A269">0.05</span> times faster than &apos;<span style="color:#A347BA">./logical-main ./crates/ruff/resources/test/cpython/ --no-cache</span>&apos;</pre>

### Default Rules -- Cache

<pre> 
<b>Summary</b>
  &apos;<span style="color:#2AA1B3">./no-logical-main ./crates/ruff/resources/test/cpython/</span>&apos; ran
<span style="color:#26A269"><b>    1.00</b></span> ± <span style="color:#26A269">0.06</span> times faster than &apos;<span style="color:#A347BA">./no-logical-pr ./crates/ruff/resources/test/cpython/</span>&apos;
<span style="color:#26A269"><b>    2.07</b></span> ± <span style="color:#26A269">0.11</span> times faster than &apos;<span style="color:#A347BA">./logical-pr ./crates/ruff/resources/test/cpython/</span>&apos;
<span style="color:#26A269"><b>    2.08</b></span> ± <span style="color:#26A269">0.12</span> times faster than &apos;<span style="color:#A347BA">./logical-main ./crates/ruff/resources/test/cpython/</span>&apos;
</pre>

### All Rules - No cache

<pre>
<b>Summary</b>
  &apos;<span style="color:#2AA1B3">./no-logical-pr ./crates/ruff/resources/test/cpython/ --no-cache --select=ALL -e</span>&apos; ran
<span style="color:#26A269"><b>    1.06</b></span> ± <span style="color:#26A269">0.03</span> times faster than &apos;<span style="color:#A347BA">./no-logical-main ./crates/ruff/resources/test/cpython/ --no-cache --select=ALL -e</span>&apos;
<span style="color:#26A269"><b>    1.07</b></span> ± <span style="color:#26A269">0.03</span> times faster than &apos;<span style="color:#A347BA">./logical-pr ./crates/ruff/resources/test/cpython/ --no-cache --select=ALL -e</span>&apos;
<span style="color:#26A269"><b>    1.24</b></span> ± <span style="color:#26A269">0.03</span> times faster than &apos;<span style="color:#A347BA">./logical-main ./crates/ruff/resources/test/cpython/ --no-cache --select=ALL -e</span>&apos;
</pre>

### All Rules - Cache
<pre><b>Summary</b>
  &apos;<span style="color:#2AA1B3">./no-logical-pr check ./crates/ruff/resources/test/cpython/ --select=ALL -e</span>&apos; ran
<span style="color:#26A269"><b>    1.08</b></span> ± <span style="color:#26A269">0.04</span> times faster than &apos;<span style="color:#A347BA">./logical-pr check ./crates/ruff/resources/test/cpython/ --select=ALL -e</span>&apos;
<span style="color:#26A269"><b>    1.10</b></span> ± <span style="color:#26A269">0.04</span> times faster than &apos;<span style="color:#A347BA">./no-logical-main check ./crates/ruff/resources/test/cpython/ --select=ALL -e</span>&apos;
<span style="color:#26A269"><b>    1.22</b></span> ± <span style="color:#26A269">0.05</span> times faster than &apos;<span style="color:#A347BA">./logical-main check ./crates/ruff/resources/test/cpython/ --select=ALL -e</span>&apos;
</pre>