@MichaReiser 2023-03-27
> ## PR Check Results
> ### Ecosystem
> 
> information_source ecosystem check **detected changes**. (+1, -1, 0 error(s))
> bokeh (+1, -1)


Using `"` is correct (which it now detects from my understanding). Source file https://github.com/bokeh/bokeh/blob/branch-3.2/src/bokeh/util/browser.py